+++ sdklib/http/__init__.py
@@ -1,11 +1,12 @@
-from sdklib.http.sdk_base import HttpSdk, HttpRequestContext
+from sdklib.http.sdk_base import HttpSdk, HttpRequestContext, generate_url_path
 from sdklib.http.response import HttpResponse
 from sdklib.http.renderers import get_renderer, url_encode
 from sdklib.util.design_pattern import Singleton
 
 
 __all__ = [
-    'HttpSdk', 'HttpResponse', 'get_renderer', 'HttpRequestContext', 'api', 'HttpRequestContextSingleton', 'url_encode'
+    'HttpSdk', 'HttpResponse', 'get_renderer', 'HttpRequestContext', 'api', 'HttpRequestContextSingleton', 'url_encode',
+    'generate_url_path'
 ]